pyproject.toml
[tool.poetry]
name = "simple-dataops-docker"
version = "0.1.0"
description = "A simple DataOps for wine recognition dataset on Docker"
authors = [
    "dongminlee94 <kid33629@gmail.com>",
    "Kimdongui <rkdqus2006@naver.com>",
]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
apache-airflow = "2.5.0"
scikit-learn = "^1.2.2"
pymongo = "^4.3.3"
pandas = "^2.0.1"
pymysql = "^1.0.3"

[tool.poetry.dev-dependencies]
black = ">=23.3.0"
pytest = ">=7.3.0"
pre-commit = ">=3.2.2"
pyright = ">=1.1.309"
ruff = ">=0.0.261"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110

[tool.isort]
profile = "black"
skip_gitignore = true

# Python Formatter
[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep-naming
    "D",  # pydocstyle
    "B",  # flakes-bugbear
    "ANN",# flakes-annotations
    "A",  # flakes-builtins
    "COM",# flakes-commas
    "T10",# flakes-debugger
    "SIM",# flakes-simplify
    "ARG",# flakes-unused-arguments
    "PLC",# pylint-convention
    "PLE",# pylint-errors
    "UP",  # pyupgrade
    "PL"  # pylint
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["I", "D", "COM", "PLC", "UP", "W292"]
ignore = ["D104", "D406", "D407", "D408", "D409", "D415", "D203", "D213", "D417", "B905", "ANN002", "ANN003", "ANN101", "ANN102", "ANN401", "ANN204", "COM812", "D107", "N812", "A003", "B008"]

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

# Same as Black.
line-length = 110
show-source = true

# Assume Python 3.10.
target-version = "py310"

pylint = {"max-args"=12, "max-statements"=60}

# Python static type checker
[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "Linux"
include = ["./src", "./docker"]
exclude = ["**/__pycache__"]
ignore = []
useLibraryCodeForTypes = true
reportMissingImports = true
reportMissingTypeStubs = false
reportOptionalSubscript = false
reportGeneralTypeIssues = false
